ConversationFunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
bs4
markdownify
requests
azure.cognitiveservices.speech
AzureOpenAI
python-dotenv
openai
pydub
pypandoc
flask
fastapi